chore(): create utils
