+++ package.json
@@ -53,7 +53,8 @@
     "rimraf": "^2.6.2",
     "semantic-release": "^15.9.17",
     "tslint": "^5.12.0",
-    "tslint-ionic-rules": "0.0.21"
+    "tslint-ionic-rules": "0.0.21",
+    "typescript-tslint-plugin": "0.3.1"
   },
   "peerDependencies": {
     "@angular-devkit/architect": ">=0.7.2",
