+++ schematics/util/index.ts
@@ -1,0 +1,13 @@
+import { strings } from '@angular-devkit/architect/node_modules/@angular-devkit/core';
+
+export function buildSelector(options: any, projectPrefix: string) {
+  let selector = strings.dasherize(options.name);
+
+  if (options.prefix) {
+    selector = `${options.prefix}-${selector}`;
+  } else if (options.prefix === undefined && projectPrefix) {
+    selector = `${projectPrefix}-${selector}`;
+  }
+
+  return selector;
+}
